Added a way to safely allow auto registration on 802.1x with MAC-AUTH on the guest SSID.
This code basically aims to unregister any node that tries MAC_AUTH if
it has been auto registered before, forcing it to reregister.

It also saves the role and unreg date to the database when doing auto
registration. Not that it would matter in this case, but it would allow
the admin to see it in the GUI.

modified:   lib/pf/node.pm
Added an is_node_auto-registered sub.

modified:   lib/pf/radius.pm
Added checks to deregister a node that has auto registered if it tries
MAC_AUTH.

modified:   lib/pf/vlan.pm
Added the unreg date and role to the nodeinfo for autoreg.